SchoolOfFish: Enhancements to the school
A Gaussian was used in the vertical axis to make the school
less of a cylinder and spread vertical fish distribution

Also, enhancements were made to school rotation and pulse
behavior, as well, as bump number of fishes to 1000.

Finally, placement of school and OrcaPod was updated to avoid colliding.